finish overview refactoring
