+++ client/src/util/stateType.ts
@@ -1,1 +1,3 @@
 export type SetState<T> = React.Dispatch<React.SetStateAction<T>>;
+
+export type UseState<T> = [T, SetState<T>];
